Manager Movies, template Manager Movies, model, Service about Movies

+++ src/main/java/com/webphim/webphim/Reponsitory/RatingMoviesRepository.java
@@ -3,5 +3,5 @@ package com.webphim.webphim.Reponsitory;
 import com.webphim.webphim.Model.RatingMovies;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RatingMoviesRepository extends JpaRepository<RatingMovies, Integer> {
+public interface RatingMoviesRepository extends JpaRepository<RatingMovies, Long> {
 }
